added trade posts
users can create posts to trade an item for an item. Backend determines whether that item is in the system yet so that when someone creates a post trading 'apple' for 'banana', and someone else creates a post trading 'banana' for 'apple', the backend is able to match the two posts as the same items. Matching still is not implemented tho